requirements.txt
numpy
pandas
scikit-learn
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6
absl-py==1.3.0
altair==4.2.0
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.4
emoji==2.2.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.8.2
Flask==2.2.2
flatbuffers==22.12.6
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.11.0
libclang==14.0.6
lxml==4.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.2.0
streamlit==1.9.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
virtualenv==20.17.1
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.14.1
zipp==3.8.0
